Added DB schema + register page

+++ StShop.DAL/Models/UserRole.cs
@@ -4,7 +4,7 @@ using System.Text;
 
 namespace StShop.DAL.Models
 {
-    public enum UserRole: int
+    public enum UserRole: byte
     {
         Customer,
         Admin
